refactor: Extract file search logic and UI from ViewFileTree into a new component.
